requirements.txt
numpy==1.26.4
openai==1.31.0
pandas==2.2.2
python-dateutil==2.9.0.post0
regex==2024.5.15
scipy==1.13.1
sortedcontainers==2.4.0
tqdm==4.66.4
